comid: implement Crypto Keys
Implement Crypto Keys as described by section 3.1.4.1.6 of
draft-ietf-rats-corim-02:

https://www.ietf.org/archive/id/draft-ietf-rats-corim-02.html#name-crypto-keys

The replaces the equivalent VerifKey construct in the existing
implementation. Crypo Keys are one of pre-defined types, allowing for
stronger validation of the underlying value.

As part of this change, the dependency on swid library is updated to the
most current version, which changes the representation of digests to use
";" instead of ":" as the separator.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>