Switch to v2 development
There is already a v1 release published to pkg.dev. This was premature
as a number of breaking changes to the API are still planned. The
underlying spec is still in development with regular drafts being
released. This code is going to change tracking the changes in the spec,
as well as likely making non-spec changes to the API as they slowly
mature.

In order to stay compliant with golang [versioning
workflow](https://go.dev/doc/modules/release-workflow), development is
witching over to v2, so that v1 can remain stable.

(note: the first full v2 release is expected to align with the
final publication of the CoRIM spec, and so will not happen for some
time. In the mean time, "release candidates" will be published to
provided milestone references for dependents)

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>