Example extension mechanism
This is a PoC for an extension mechanism that would allow calling code
to supply extension implementations for sockets defined by the CoRIM
spec.

This is done by embedding an Extensions object at socket locations. This
object wraps an Interface implementations of which are user-supplied.
Extensions provides uniform accessor mechanism for the user extensions
without them needing to implement themselves. A custom serialisation
mechanism is used to ensure correct CBOR is generated for the embedded
structs.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>