refactor!: merge key triple types into KeyTriple
Replace DevIdentityKey and AttestVerifKey with KeyTriple type that is
used for both.

The two types were identical. They were the same structure being used
for two purposes. The purposes are already reflected in the field names
within Triples structure, and there is no need to also include it in the
type name, which only leads to unnecessary code duplication.

Additionally, the old names were misleading as they implied that the
types represented a single key, whereas they in fact contained a
collection of keys.

Finally, Alias []KeyTriple to KeyTriples type. This aligns with how
collections are handled elsewhere and provides a more consistent API.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>