fix!: serialize tags as tags rather than bstr
The CoRIM spec defines the tags field of corim-map as

      &(tags: 1) => [ + $concise-tag-type-choice ]

where

    $concise-tag-type-choice /= tagged-concise-swid-tag
    $concise-tag-type-choice /= tagged-concise-mid-tag
    $concise-tag-type-choice /= tagged-concise-tl-tag

i.e. tags field is supposed to be an array of one or more tagged
objects.

Our UnsignedCorim.Tags was defined as []Tag, where Tag was an alias of
[]byte. It is an array of serialized tag objects. However, since it is
essentially, an array of byte strings, it was serialized as such,
resulting in

    [ + bstr .cbor $concise-tag-type-choice ]

instead of the expected

   [ + $concise-tag-type-choice ]

To get around this, redefined UnsignedCorim.Tags to be []cbor.RawMessage
(this is also an alias of []bytes, but the cbor library will not try to
encode it as bstr, instead writing the byte array into the serialized
output verbatim).

BREAKING CHANGE: tags (CoMID, CoSWID, etc) are now serialized as cbor
tagged maps, rather than cbor bstr's. API-wise, the Tag type has been
removed and replaced with cbor.RawMessage.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>