fix!: serialize tags as tagged bstr
The CoRIM spec defines the tags field of corim-map as

      &(tags: 1) => [ + $concise-tag-type-choice ]

where

    $concise-tag-type-choice /= tagged-concise-swid-tag
    $concise-tag-type-choice /= tagged-concise-mid-tag
    $concise-tag-type-choice /= tagged-concise-tl-tag

and

    tagged-concise-swid-tag = #6.505(bytes .cbor concise-swid-tag)
    tagged-concise-mid-tag = #6.506(bytes .cbor concise-mid-tag)
    tagged-concise-tl-tag = #6.508(bytes .cbor concise-tl-tag)

i.e. tags field is supposed to be an array of one or more tagged
byte string, which are themselves CBOR encodings of the corresponding
struct.

Our UnsignedCorim.Tags was defined as []Tag, where Tag was an alias of
[]byte. It is an array of serialized tag objects. However, since it is
essentially, an array of byte strings, it was serialized as such,
resulting in

    [ + bstr .cbor TAG(OBJECT) ]

instead of the expected

   [ + TAG(bstr .cbor OBJECT) ]

i.e. tag ended up being part of the bstr, where as it ought to have been
wrapping the bstr.

This commit changes the definition of Tag to a struct that associates a
tag number with a []byte (it is similar to cbor.Tag except the content
is always a []byte rather than any). In CBOR, this type is serialized
exactly as a cbor.Tag; in JSON, this includes the tag in the
base64-encoded data.

BREAKING CHANGE: tags (CoMID, CoSWID, etc) are now serialized as tagged
bstr's, rather than plain bstrs (that contained the tag). API-wise, the
Tag type has been changed to a struct that associates a tag number with
a []byte.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>