fix(comid): AddDigests returned nil on error
Measurement.AddDigests follows the general setter pattern in the API by
only returning a pointer to the modified structure, and not an error to
allow chaining.

However, Digests.AddDigests returned nil (rather than pointer to the
collection) if the provided inputs were in valid. This resulted in
potential panics if someone tries to chain another setter after that (as
our examples show they should be able to do).

Modify AddDigests to always return the pointer to the modified
structure. It no longer validates the inputs, and instead relies on
subsequent calls to Valid() to catch malformed digests.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>