chore(comid): simplify RawValue.SetBytes implementation
Until now, SetBytes would create TaggedBytes by calling NewBytes() on
its input, and returning nil instead of a pointer to itself on error.

This _would_ be a problem for chaining, except an error would never
occur, as NewBytes() only errors if it is given something it doesn't
know how to convert to a []byte. And RawValue.SetBytes always gave it
[]bytes.

To simplify things, and remove potential confusion from anyone else
reading the code, replace the call to NewBytes() with a direct cast of
the input []byte to TaggedBytes, removing the need for superfluous error
handling.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>