feat: unknown field caching when decoding
Allow caching of unknown fields during PopulateStructFrom(CBOR|JSON).
The cache is then used during SerializeStructTo(CBOR|JSON). This ensures
that any map entries in the source data that do not correspond to target
struct fields are preserved across decode/encode cycle.

This feature is enabled by adding a field to a struct tagged with
`field-cache:""`. This field must be of type map[string]any. When
PopulateStructFrom* functions encounter input entries that do not
correspond to a field in the target struct, they will be added to the
field-cache map instead. Analogously, when SerializeStructTo* functions
see a field-cache map, they will add its entries to the output.

This scheme has some (hopefully, obvious) limitations:

- field-cache field's tag must also contain `cbor:"-" json:"-"` to make
  sure that the field itself will be ignored by serializers.
- if a struct is unmarshaled from JSON, any unknown field names
  must be "stringified" integers, otherwise it is impossible to obtain
  the corresponding CBOR code point mapping for the name.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>