fix: enforce lexicographic ordering when encoding CBOR
Up to this point, we only had deterministic encoding for structs. Fields
were encoded in the order they appeared.

This fix ensures that fields are encoded in lexicographic order as
required by CDE spec:

    https://www.ietf.org/archive/id/draft-ietf-cbor-cde-13.html#name-the-lexicographic-map-sorti

(Note: we typically define fields in the order of their code points, so
in practice, we have been _mostly_ compliant with CDE; but this was not
guaranteed, and deviations were possible, especially when extensions are
involved.)

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>