feat: implement extension caching
Extensions objects will now cache any extensions they don't recognize
(that don't correspond to a field inside their registered IMapValue)
when deserializing values.

This means CORIMs with extensions remain stable when deserialized and
then re-serialized using structs without registered extensions.

Cached extensions are only used during marshalling, and will not be
returned when calling the Get* methods. However they can be accessed
directly via Extensions.Cached field.

When an IMapValue struct is registered, cached values are scanned, and
the new struct is populated with cached values, which are then removed
form the cache.

When registering a new struct when there is an existing IMapValue, the
non-zero-value fields of the old IMapValue are stored in the cache.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>